fature-manage-fix-migration

+++ app/Http/Controllers/GroupController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\UpdateGroupRequest;
 use App\Repositories\Interfaces\GroupRepositoryInterface;
 use Illuminate\Http\Request;
 use App\Models\Group;
+use App\Models\Role;
+use Illuminate\Support\Facades\Log;
 
 class GroupController extends Controller
 {
@@ -20,6 +22,7 @@ class GroupController extends Controller
     }
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Group::class);
         $items = $this->GroupService->all($request);
         return view('groups.index', compact('items'));
     }
@@ -29,6 +32,7 @@ class GroupController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Group::class);
         return view('groups.create');
     }
 
@@ -38,7 +42,7 @@ class GroupController extends Controller
     public function store(StoreGroupRequest $request)
     {
 
-        $data = $request->except(['_token','_method']);
+        $data = $request->except(['_token', '_method']);
         $this->GroupService->store($data);
         return redirect()->route('groups.index');
     }
@@ -46,19 +50,13 @@ class GroupController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
-    {
-        $Groups = Group::get();
-        $items = $this->GroupService->find($id);
-        // dd($Users);
-        return view('groups.show', compact('items', 'Groups'));
-    }
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(string $id)
     {
+        $this->authorize('update', Group::class);
         $item = $this->GroupService->find($id);
         return view('groups.edit', compact('item'));
     }
@@ -66,11 +64,12 @@ class GroupController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateGroupRequest $request,$id)
+    public function update(UpdateGroupRequest $request, $id)
     {
-        $data = $request->except(['_token','_method']);
-        $this->GroupService->update($id,$data);
-            return redirect()->route('groups.index');
+
+        $data = $request->except(['_token', '_method']);
+        $this->GroupService->update($id, $data);
+        return redirect()->route('groups.index');
     }
 
     /**
@@ -78,7 +77,33 @@ class GroupController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete', Group::class);
         $this->GroupService->destroy($id);
         return redirect()->route('groups.index');
+    }
+    public function show(string $id)
+    {
+        $group =  $this->GroupService->show($id);
+        $roles = Role::get();
+        $group = $this->GroupService->find($id);
+        $params = [
+            'group' => $group,
+            'roles'   => $roles
+        ];
+
+        return view('groups.show', $params);
+    }
+
+
+    public function group_detail(Request $request, $id)
+    {
+
+        try {
+            $this->GroupService->group_detail($id, $request);
+            return redirect()->route('groups.index');
+        } catch (\exception $e) {
+            Log::error($e->getMessage());
+            return redirect()->route('groups.index');
+        }
     }
 }
